- fixup the array classes to avoid breaking bug(s) due to different type of container - fixup superweapon type stuffs - fix some SW cannot fire due to parsing problem - fix problem with asmjit hooks beign recursively called causing game to randomly crash - fix TActionExtData numbering error - fix game freeze when exit - backport core parts of TeamClass internal functions - update libs - remove some unused libs to reduce compile time - sync some to develop

+++ src/Ext/TechnoType/Body.cpp
@@ -2027,7 +2027,7 @@ bool TechnoTypeExtData::LoadFromINI(CCINIClass* pINI, bool parseFailAddr)
 
 		this->SprayOffsets.resize(defaultSprayOffsets.size());
 
-		for (int i = 0; i < defaultSprayOffsets.size(); ++i) {
+		for (size_t i = 0; i < defaultSprayOffsets.size(); ++i) {
 			this->SprayOffsets[i] = defaultSprayOffsets[i];
 		}
